model/requirements.txt
MarkupSafe
filelock
fsspec
jinja2
mpmath
networkx
pillow
sympy
torch
torchaudio
torchvision
pandas
kaggle
matplotlib
transformers
scikit-learn
nltk
temporalio
dill
pytesseract
minio